requirements.txt
fastapi==0.121.0
openai==2.7.1
pydantic==2.12.4
python-dotenv==1.2.1
uvicorn==0.31.0